requirements.txt
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
cycler==0.11.0
demjson==1.6
h11==0.12.0
idna==3.3
jtutils==0.0.8
kiwisolver==1.3.2
leven==1.0.4
matplotlib==3.4.3
mysql-connector==2.2.9
mysql-connector-python==8.0.27
nose==1.3.7
numpy==1.21.3
outcome==1.1.0
pandas==1.3.4
Pillow==8.4.0
protobuf==3.19.1
pycparser==2.20
pyOpenSSL==21.0.0
pyparsing==3.0.4
python-csv==0.0.13
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
selenium==4.0.0b4
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
trio==0.19.0
trio-websocket==0.9.2
undetected-chromedriver==3.0.3
urllib3==1.26.7
websockets==10.0
wsproto==1.0.0
xlrd==2.0.1
xmltodict==0.12.0
